Update Make Square and Bitboard operators commutative
As Stockfish developers, we aim to make our code as legible and as close
to simple English as possible. However, one of the more notable
exceptions
to this rule concerns operations between Squares and Bitboards.

Prior to this pull request, AND, OR, and XOR were only defined when the
Bitboard was the first operand, and the Square the second. For example,
for a Bitboard b and Square s, "b & s" would be valid but "s & b" would
not.
This conflicts with natural reasoning about logical operators, both
mathematically and intuitively, which says that logical operators should
commute.

More dangerously, however, both Square and Bitboard are defined as
integers
"under the hood." As a result, code like "s & b" would still compile and
give
reasonable bench values. This trap occasionally ensnares even
experienced
Stockfish developers, but it is especially dangerous for new developers
not
aware of this peculiarity. Because there is no compilation or runtime
error,
and a reasonable bench, only a close review by approvers can spot this
error
when a test has been submitted--and many times, these bugs have slipped
past
review. This is by far the most common logical error on Fishtest, and
has
wasted uncountable STC games over the years.

However, it can be fixed by adding three non-functional lines of code.
In this
patch, we define the operators when the operands are provided in the
opposite
order, i.e., we make AND, OR, and XOR commutative for Bitboards and
Squares.
Because these are inline methods and implemented identically, the
executable
does not change at all.

This patch has the small side-effect of requiring Squares to be
explicitly
cast to integers before AND, OR, or XOR with integers. This is only
performed
twice in Stockfish's source code, and again does not change the
executable at
all (since Square is an enum defined as an integer anyway).

For demonstration purposes, this pull request also inverts the order of
one AND
and one OR, to show that neither the bench nor the executable change.
(This
change can be removed before merging, if preferred.)

I hope that this pull request significantly lowers the barrier-of-entry
for new
developer to join the Stockfish project. I also hope that this change
will improve
our efficiency in using our generous CPU donors' machines, since it will
remove
one of the most common causes of buggy tests.

Following helpful review and comments by Michael Stembera (@mstembera),
we add
a further clean-up by implementing OR for two Squares, to anticipate
additional
traps developers may encounter and handle them cleanly.

Closes #2387

No functional change.
master (#1130)
@31m059
@snicolet
31m059 authored