Sequencing tweak in tbprobe()
Followup of "issue" #2372, which was in fact a small speed-up
proposal by user @d3vv for the probing code of tablebases.

See comments on this issue where it was proven by Alin Savard that the
proposed change is more efficient on average than master on all type of
sequences it will usually be called.

Note that on gcc 4.3, this will produce a bogus warning which was solved
with ulterior gcc versions:
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=43949

Closes #2372
Closes #2379

Non functional change
master
@Rocky640
@snicolet
Rocky640 authored